Add files via upload
Neither of these are complete and the first one is literally just our first assignment with my attempt at new functions near the bottom. The other one was my trying to implement a Gini formula and failing 

As for how I was trying to contrust Gini, see https://stackoverflow.com/questions/31321810/gini-coefficient-in-julia-efficient-and-accurate-code and https://zhiyzuo.github.io/Plot-Lorenz/ (in Python).